requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
Django==4.2.1
django-phonenumber-field==7.1.0
django-phonenumbers==1.0.1
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
phonenumbers==8.13.13
psycopg2==2.9.6
PyJWT==2.8.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
secure-smtplib==0.1.1
sqlparse==0.4.4
twilio==8.5.0
typing_extensions==4.6.3
urllib3==2.0.4
yarl==1.9.2
